refactor: change the ui types to utils
- change the folder of some swap card tokens types to utils
