+++ components/02-molecules/OfferSummary.tsx
@@ -1,9 +1,10 @@
 import { ENSAvatar, ENSAvatarSize, PersonIcon } from "@/components/01-atoms";
-import { TokenCardStyleType, TokensList } from "@/components/02-molecules";
+import { TokensList } from "@/components/02-molecules";
 import { ForWhom } from "@/lib/client/constants";
 import { SwapContext } from "@/lib/client/contexts";
 import { useAuthenticatedUser } from "@/lib/client/hooks/useAuthenticatedUser";
 import { useEnsData } from "@/lib/client/hooks/useENSData";
+import { TokenCardStyleType } from "@/lib/client/ui-utils";
 import { useContext } from "react";
 
 export const OfferSummary = ({ variant }: { variant: ForWhom }) => {
