+++ components/02-molecules/TokenCard.tsx
@@ -8,7 +8,12 @@ import {
   Token,
   TokenType,
 } from "@/lib/shared/types";
-import { getTokenName } from "@/lib/client/ui-utils";
+import {
+  TokenCardActionType,
+  TokenCardStyleType,
+  TokenSizeClassNames,
+  getTokenName,
+} from "@/lib/client/ui-utils";
 import { SwapContext } from "@/lib/client/contexts";
 import React, { useContext, useEffect, useState } from "react";
 import cc from "classcat";
@@ -33,33 +38,12 @@ interface TokenCardProps {
   styleType?: StyleVariant;
 }
 
-export enum TokenCardActionType {
-  "SELECT_TOKEN_FOR_SWAP",
-  "APPROVE_TOKEN_SWAP",
-  "SHOW_NFT_DETAILS",
-  "NO_ACTION",
-}
-
-export enum TokenCardStyleType {
-  SMALL = "small",
-  NORMAL = "normal",
-  MEDIUM = "medium",
-  LARGE = "large",
-}
-
 type StyleVariant =
   | TokenCardStyleType
   | "small"
   | "normal"
   | "medium"
   | "large";
-
-export const TokenSizeClassNames = {
-  [TokenCardStyleType.SMALL]: "card-token-small",
-  [TokenCardStyleType.NORMAL]: "card-token-normal",
-  [TokenCardStyleType.MEDIUM]: "card-token-medium",
-  [TokenCardStyleType.LARGE]: "card-token-large",
-};
 
 /**
  * TokenCard Component
